Tired of working around siteminder bad design.
The story is if I embed with ?skin=0 there are no scrollbars for desktop
browsers (works fine at mobile dimensions). If we embed without a skin
the scrollbars appear but then we get all the "skin" stuff we don't
need and our brand header is gone which defeats the point.

As the iframe is from another domain I cannot script it due to cross
origin policies and therefore cannot stretch the containing element to
the height of the iframe content.